fix watcher bug

+++ deploy.sh
@@ -8,6 +8,13 @@ proj_name=$(cat project_name)
 
 sh compile.sh $@
 
-rm -rf "$tomcat_home"webapps/"$proj_name"
+command_exists () {
+    type "$1" &> /dev/null ;
+}
+
+if ! command_exists open; then
+    rm -rf "$tomcat_home"webapps/"$proj_name"
+fi
+
 echo cp -rf release/"$proj_name"/ \""$tomcat_home"webapps/"$proj_name"\"
 cp -rf release/"$proj_name"/ "$tomcat_home"webapps/"$proj_name"
